# [Perf] Cuda Kernel for Int8 Per Token Group Quant