## Purpose

Add support for int8 per-token group quant CUDA kernel

Follow up pr for https://github.com/vllm-project/vllm/pull/21083

## Test

Unit test:

```bash
(wentao) wentao@H100-GPU17:~/vllm/tests/kernels/quantization$ pytest test_per_token_group_quant.py
================================================================= test session starts =================================================================
platform linux -- Python 3.12.11, pytest-8.4.1, pluggy-1.6.0
rootdir: /home/wentao/vllm
configfile: pyproject.toml
plugins: anyio-4.9.0
collected 30 items                                                                                                                                    

test_per_token_group_quant.py ..............................                                                                                    [100%]

================================================================= 30 passed in 6.38s ==================================================================
```

Benchmark: **15x faster compared with triton**

```bash
(wentao) wentao@H100-GPU17:~/vllm/benchmarks/kernels$ python benchmark_per_token_group_quant.py 
Configuration                                           |  CUDA (ms)   |  Triton (ms)  | Speed-up
-------------------------------------------------------------------------------------------------
shape=(32, 128)  gs=64   col_major=0      ue8m0=0      dtype=fp8 | CUDA   0.011 ms  | Triton   0.169 ms  | speed-up ×15.73
shape=(32, 128)  gs=64   col_major=0      ue8m0=1      dtype=fp8 | CUDA   0.011 ms  | Triton   0.159 ms  | speed-up ×14.99
shape=(32, 128)  gs=64   col_major=1      ue8m0=0      dtype=fp8 | CUDA   0.012 ms  | Triton   0.168 ms  | speed-up ×13.83
shape=(32, 128)  gs=64   col_major=1      ue8m0=1      dtype=fp8 | CUDA   0.012 ms  | Triton   0.167 ms  | speed-up ×13.86
shape=(32, 128)  gs=128  col_major=0      ue8m0=0      dtype=fp8 | CUDA   0.011 ms  | Triton   0.159 ms  | speed-up ×15.13
shape=(32, 128)  gs=128  col_major=0      ue8m0=1      dtype=fp8 | CUDA   0.010 ms  | Triton   0.158 ms  | speed-up ×15.22
shape=(32, 128)  gs=128  col_major=1      ue8m0=0      dtype=fp8 | CUDA   0.012 ms  | Triton   0.169 ms  | speed-up ×13.84
shape=(32, 128)  gs=128  col_major=1      ue8m0=1      dtype=fp8 | CUDA   0.012 ms  | Triton   0.167 ms  | speed-up ×13.94
shape=(64, 256)  gs=64   col_major=0      ue8m0=0      dtype=fp8 | CUDA   0.010 ms  | Triton   0.156 ms  | speed-up ×14.97
shape=(64, 256)  gs=64   col_major=0      ue8m0=1      dtype=fp8 | CUDA   0.010 ms  | Triton   0.156 ms  | speed-up ×14.88
shape=(64, 256)  gs=64   col_major=1      ue8m0=0      dtype=fp8 | CUDA   0.012 ms  | Triton   0.167 ms  | speed-up ×13.59
shape=(64, 256)  gs=64   col_major=1      ue8m0=1      dtype=fp8 | CUDA   0.012 ms  | Triton   0.167 ms  | speed-up ×13.63
shape=(64, 256)  gs=128  col_major=0      ue8m0=0      dtype=fp8 | CUDA   0.011 ms  | Triton   0.158 ms  | speed-up ×14.90
shape=(64, 256)  gs=128  col_major=0      ue8m0=1      dtype=fp8 | CUDA   0.010 ms  | Triton   0.158 ms  | speed-up ×15.10
shape=(64, 256)  gs=128  col_major=1      ue8m0=0      dtype=fp8 | CUDA   0.012 ms  | Triton   0.167 ms  | speed-up ×13.79
shape=(64, 256)  gs=128  col_major=1      ue8m0=1      dtype=fp8 | CUDA   0.012 ms  | Triton   0.168 ms  | speed-up ×13.87
shape=(16, 512)  gs=64   col_major=0      ue8m0=0      dtype=fp8 | CUDA   0.011 ms  | Triton   0.157 ms  | speed-up ×14.77
shape=(16, 512)  gs=64   col_major=0      ue8m0=1      dtype=fp8 | CUDA   0.011 ms  | Triton   0.157 ms  | speed-up ×14.86
shape=(16, 512)  gs=64   col_major=1      ue8m0=0      dtype=fp8 | CUDA   0.012 ms  | Triton   0.167 ms  | speed-up ×13.82
shape=(16, 512)  gs=64   col_major=1      ue8m0=1      dtype=fp8 | CUDA   0.012 ms  | Triton   0.168 ms  | speed-up ×13.84
shape=(16, 512)  gs=128  col_major=0      ue8m0=0      dtype=fp8 | CUDA   0.010 ms  | Triton   0.158 ms  | speed-up ×15.06
shape=(16, 512)  gs=128  col_major=0      ue8m0=1      dtype=fp8 | CUDA   0.011 ms  | Triton   0.158 ms  | speed-up ×14.97
shape=(16, 512)  gs=128  col_major=1      ue8m0=0      dtype=fp8 | CUDA   0.012 ms  | Triton   0.168 ms  | speed-up ×13.80
shape=(16, 512)  gs=128  col_major=1      ue8m0=1      dtype=fp8 | CUDA   0.012 ms  | Triton   0.167 ms  | speed-up ×13.71
```

Int8

```bash
shape=(32, 128)  gs=64   col_major=0      ue8m0=0      dtype=int8 | CUDA   0.010 ms  | Triton   0.156 ms  | speed-up ×15.47
shape=(32, 128)  gs=128  col_major=0      ue8m0=0      dtype=int8 | CUDA   0.010 ms  | Triton   0.158 ms  | speed-up ×15.87
shape=(64, 256)  gs=64   col_major=0      ue8m0=0      dtype=int8 | CUDA   0.010 ms  | Triton   0.155 ms  | speed-up ×15.33
shape=(64, 256)  gs=128  col_major=0      ue8m0=0      dtype=int8 | CUDA   0.010 ms  | Triton   0.155 ms  | speed-up ×15.53
shape=(16, 512)  gs=64   col_major=0      ue8m0=0      dtype=int8 | CUDA   0.010 ms  | Triton   0.156 ms  | speed-up ×15.57
shape=(16, 512)  gs=128  col_major=0      ue8m0=0      dtype=int8 | CUDA   0.010 ms  | Triton   0.156 ms  | speed-up ×15.42
```

I don't know which model to test on, please let me know. But basically I think it would give similar e2e performance like fp8 (6% improvement for throughput)