@huydhn 2025-07-26
I think this change is breaking ROCm https://github.com/pytorch/pytorch-integration-testing/actions/runs/16534012953/job/46765135029#step:14:1136.  It's a bit hard to see the failure because it's a warning, not a crash.

```
WARNING 07-26 00:32:16 [rocm.py:34] Failed to import from vllm._C with ImportError('/usr/local/lib/python3.12/dist-packages/vllm/_C.abi3.so: undefined symbol: _Z26per_token_group_quant_8bitRKN2at6TensorERS0_S3_ldddb')
WARNING 07-26 00:32:17 [_custom_ops.py:20] Failed to import from vllm._C with ImportError('/usr/local/lib/python3.12/dist-packages/vllm/_C.abi3.so: undefined symbol: _Z26per_token_group_quant_8bitRKN2at6TensorERS0_S3_ldddb')
```

cc @simon-mo @mgoin @yewentao256