@huydhn 2025-07-26
AMD tests are failing after this lands https://buildkite.com/vllm/ci/builds/24999#0198440f-053a-463d-a191-3d7bdbe834db.  No build failure though. cc @houseroad